ComputePowers.samplerate now cast to float; necessary for sys2 biomarkers

+++ ReportTasks/RamTaskMethods.py
@@ -67,7 +67,7 @@ def compute_powers(events,monopolar_channels,bipolar_pairs,
     print 'Total time elapsed: {}'.format(time.time()-tic)
     # print 'Time spent on wavelet filter: {}'.format(filter_time)
     if ComputePowers is not None:
-        ComputePowers.samplerate = eeg['samplerate'].data
+        ComputePowers.samplerate = eeg['samplerate'].data.astype(np.float)
     return pow_mat,events
 
 """======================================== Classifier Functions =================================================== """
